Children threads now also report signal termination status
Before this commit, signal termination status was only
available to the parent thread.  Now it is available to both
the children and the parent thread, so that children would
report the signal by which they were terminated (SIGINT,
currently) and then execute the default action for SIGINT.

The parent's behavior of picking up the child status, and
reporting it, including signal termination of the child is
not changed.

Signed-off-by: Luben Tuikov <ltuikov@yahoo.com>